Using tuple comparisons in composite keys
* Some essential context here:
* Our ACE algorithm splits the pkey into ranges and sends them off to
multiprocessing workers. So, if there are 1000 ids as pkeys,
with block size = 100, our workers get [(1,100), (101, 200)...]

* In case of a composite primary key, how do we reliably
get block ranges for table comparisons? There are some issues
when we use something like (id, date) > (10, '2021-09-21')

* However, we *cannot* split up the composite key into different where
clauses, i.e., id > 10 AND date > '2021-09-21'. This would yield
incorrect results.

* Therefore, we stick to doing tuple comparisons and we'd like to
caution the user that the choice of the first key in the composite
key is critical -- not just for their own application, but for ace
modules as well. In simple terms, ACE is trusting the user to have
sensisble composite keys; otherwise, it may not yield correct results.