Implement Ultra High Availability (HA) Support Using Patroni and etcd.
This commit introduces support for Ultra High Availability (HA) using Patroni
and etcd. Changes include modifications to service scripts for Patroni and
etcd, the addition of Python scripts patroni.py and etcd.py for installation and
management, and removing the deprecated patron-cluster.py script. Ultra HA
functionality is now integrated into the pgedge cluster command to enhance
the reliability and resilience of pgEdge.