Added support for Spock version 4 in add-node
Implemented support for Spock version 4 in the add-node command. There are two
major differences between Spock version >= 4 and < 4. Spock version 4 uses the
ALTER command to change the cluster to read-only mode, whereas versions < 4
use an SQL function for this purpose. Additionally, in versions < 4, read-only
mode does not allow the creation of nodes and subscriptions, while in versions
>= 4, superusers are permitted to create these. Consequently, the command
sequence for add-node differs between versions >= 4 and < 4.