Fixed issue where replica nodes weren't starting.
This was being caused by track_commit_timestamps not being enabled
while spock.conflict_resolution was set to last_update_wins. Since
this is such a fundamental configuration setting, I've placed it in
the DCS bootstrap section of the Patroni config, so it's applied to
all nodes.

This also required pushing the nesting of the postgres block down one
level, as it's supposed to reside under dcs rather than bootstrap.